backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.6.0
openai==1.37.0
